Experimental support for on-expiration messaging for strings
With this change, the Redis server will dispatch a 0MQ message whenever
a key (of simple string type) expires. This is useful when using
Redis as a persistent, reliable hot cache and the data needs to be
expired into a cold storage solution.

0MQ was chosen for simplicity, low overhead and high message throughput
so that Redis isn't slowed down significantly.

Conflicts:

	src/Makefile
	src/config.c

Conflicts:
	redis.conf
	src/Makefile
	src/config.c