First stab at implementing expire looping
In a nutshell, this feature will enable stuffing a hash back into Redis
when it expires, while resetting expiration and keeping track of how
often it has done the expire loop. Eventually, when that counter hits a
configurable limit, it won't be reinserted.

This is useful as a oh-my-god-is-this-painful workaround for both
unreliable messaging and the
this-key-is-in-flight-and-not-available-from-anywhere race condition.
The worker processes that pick up the expired messages are expected
to issue an explicit delete from Redis when they've successfully
inserted the record into the long term storage. That breaks the cycle.

Conflicts:
	src/db.c